requirements.txt
setuptools
numpy
bob.core
bob.extension
bob.ip.base
bob.ip.color
bob.io.image
bob.pad.base
bob.pad.face
bob.learn.pytorch
timm==0.3.4
